requirements.txt
docker==4.1.0
python-dxf==7.5.3
requests==2.22.0